STEP 16. Move out stores.
